requirements.txt
streamlit==1.26.0
pandas==2.1.1
scikit-learn==1.3.0
pymongo==4.5.0
pyrebase4
